Moves Vegetation source and header files to /src and /include
