+++ src/Vegetation.cpp
@@ -16,11 +16,11 @@
  *
  */
 
-#include "Vegetation.h"
+#include "../include/Vegetation.h"
 
-#include "../TEMUtilityFunctions.h"
+#include "TEMUtilityFunctions.h"
 
-#include "../TEMLogger.h"
+#include "TEMLogger.h"
 extern src::severity_logger< severity_level > glg;
 
 Vegetation::Vegetation() {
